pyproject.toml
[tool.poetry]
name = "hunters-jukebox"
version = "0.1.0"
description = ""
authors = ["Bryce Clark <bryce@lowquality.dev>"]
readme = "README.md"
packages = [{include = "hunters_jukebox"}]

[tool.poetry.dependencies]
python = "^3.11"
discord-py = {extras = ["voice"], version = "^2.2.3"}
pynacl = "^1.5.0"
youtube-dl = { git = "https://github.com/ytdl-org/youtube-dl.git", branch="master" }
yt-dlp = "^2023.3.4"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
